refactor : 지켜지지않는폴더컨벤션을수정했습니다
