+++ src/App.tsx
@@ -3,8 +3,8 @@ import Home from './pages/home'
 import PostCreation from './pages/post-creation'
 import Profile from './pages/profile'
 import Search from './pages/search'
-import PostEdit from './pages/PostEdit'
-import PostDetail from './pages/PostDetail'
+import PostEdit from './pages/post-edit'
+import PostDetail from './pages/post-detail'
 import LoginForm from './pages/login/LoginForm'
 import SignUpForm from './pages/sign-up/SignupForm'
